+++ administrator/components/com_loginguard/src/Extension/LoginGuardComponent.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Joomla\Component\LoginGuard\Administrator\Extension;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\MVCComponent;
+
+final class LoginGuardComponent extends MVCComponent
+{
+}